### 2016-05-11 #17 @whedon
Submission: Application Skeleton: Generating Synthetic Applications for Infrastructure Research
**Submitting author:** @zhaozhang (<a href="http://orcid.org/0000-0001-5921-0035">Zhao Zhang</a>)
**Repository:** <a href="https://github.com/applicationskeleton/Skeleton" target ="_blank">https://github.com/applicationskeleton/Skeleton</a>
**Version:** v1.2
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.13750" target ="_blank">http://dx.doi.org/10.5281/zenodo.13750</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.13750">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [paper.pdf](https://github.com/openjournals/joss-reviews/files/259667/paper.pdf)
