### 2016-05-11 #17 @krother
- References: The article "Pegasus: A framework ..." has a doi that is not listed in the paper. I did not check the other doi-less articles.
- Paper: In my opinion the first paragraph of the summary deserves improvement. For instance, I find it hard to understand that Skeleton is about *distributed* applications.
