### 2016-05-11 #17 @krother
## Recommendation:
I recommend publication after a minor review of the *statement of need* both in the README and the paper. An double-check of the doi-less references would be great.

## General checks
all checked

## Functionality
all checked

- I checked the simple example from the documentaion: it worked out of the box.
- I did not check the Pegasus / Swift examples.

## Documentation

- Statement of need: needs improvement. I did not find a description of the target audience. I do not fully understand what the program is useful for.
- Installation instructions: accurate, although the installation is not automatic. I think a plain text format for the documentation would be much easier to access than the .pptx (and would look better on LibreOffice, too). Copy-pasting from pptx file did not work for this command:

    bin/aimes-skeleton-generate -i src/aimes/skeleton/sample-input/bag.input -m shell

- Example Usage: OK
- Functionality documentation: Yes, I find the manual clearly written and detailed enough. The figures are very illustrative and help understanding. There are some minor English grammar issues but they don't impede readability.
- Automated tests: No automated tests, but a precise list of manual steps and an accurate description of an use case and the expected result.
- Community guidelines: There is an invitation link to the mailing list, but not more. I am reluctant to check point 1) "guidelines for third parties wishing to contribute"

## Software paper:
- Authors: OK
- Statement of need: In my opinion the first paragraph of the summary deserves improvement. For instance, I find it hard to understand that Skeleton is about distributed applications.
- References: The article *"Pegasus: A framework ..."* has a doi that is not listed in the paper. I did not check the other doi-less articles.
