### 2016-05-11 3bedeef
added 3 DOIs
