+++ paper.bib
@@ -4,7 +4,8 @@
   booktitle={IEEE Congress on Services},
   pages={199--206},
   year={2007},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/SERVICES.2007.63},
 }
 
 @article{SWIFT09,
@@ -36,7 +37,8 @@
   volume = "13",
   number = "3",
   pages = "219--237",
-  year = "2005"
+  year = "2005",
+  doi={10.1155/2005/128026},
 }
 
 @incollection{PEGASUS04,
@@ -50,6 +52,7 @@
   pages = {131-140},
   volume = {3165},
   year = {2004}
+  doi={10.1007/978-3-540-28642-4_2},
 }
 
 @article{AIMES15,
