### 2016-05-11 #17 @danielskatz
Thanks @krother.

I've added the DOIs for the 3 papers that have DOIs now and will work on the other issues.
