+++ paper.md
@@ -31,10 +31,10 @@ bibliography: paper.bib
 ---
 
 # Summary
-Application Skeleton is a simple and powerful tool to build simplified synthetic applications with runtime and I/O close to that of the real applications. It allows computer scientists who need to use applications to 
-verify the effectiveness of new system designs to bypass obstacles that are often encountered when accessing and building real applications. Using the applications generated by Application Skeleton guarantees that the CS systems' effectiveness on synthetic applications will apply to the real applications.
+Application Skeleton is a simple and powerful tool to build simplified synthetic applications with runtime and I/O close to that of the real applications. It is intended for computer scientists who need to use applications to 
+verify the effectiveness of new system designs, to bypass obstacles that are often encountered when accessing and building real applications. Using the applications generated by Application Skeleton guarantees that the CS systems' effectiveness on synthetic applications will apply to the real applications.
 
-Application Skeleton can generate bag-of-task, (iterative) map-reduce, and (iterative) multistage workflow applications. These applications are represented as a set of tasks, a set of input files, and a set of dependencies.  The generated applications are compatible with workflow system such as Swift [@SWIFT07, @SWIFT09, @SWIFT11] and Pegasus [@PEGASUS04, @PEGASUS05], as well as the ubiquitous UNIX shell.
+Application Skeleton can generate bag-of-task, (iterative) map-reduce, and (iterative) multistage workflow applications. These applications are represented as a set of tasks, a set of input files, and a set of dependencies.  These applications can be generalized as many-task applications, and once created, can be run on single-core, single-node, multi-core, or multi-node (distributed or parallel) computers, depending on what workflow system is used to run them.  The generated applications are compatible with workflow system such as Swift [@SWIFT07, @SWIFT09, @SWIFT11] and Pegasus [@PEGASUS04, @PEGASUS05], as well as the ubiquitous UNIX shell.
 The application can also be created as a generic JSON object that can be used by other systems such as the AIMES [@AIMES15] middleware.
 
 # References
