### 2016-05-11 #17 @danielskatz
regarding the statement of need, and your comment about distributed applications, I've changed the first paragraph to try to make the need/use of the software more clear, and changed the second paragraph to explain more about what types of applications can be built.  Please let me know if this seems ok, or if you think more changes here are needed.
