### 2016-05-11 #17 @krother
Yes it's definitely better now. I am not entirely sure what to do with the word 'application' (I think mobile app, web app, Python app.. ). I understand for someone developing distributed systems the term translates to 'the program that does the work'. Please correct me if I'm on an entirely wrong track.
