### 2016-05-11 6474dfe
adding info about support and contributions
