### 2016-05-11 ec1e2bc
adding info about who might use this software in README
