+++ README.md
@@ -4,15 +4,19 @@
 Skeleton
 ========
 
-Application Skeleton is a tool to generate skeleton applications --- easy-to-program, easy-to-run applications --- that mimic a real applications' parallel or distributed performance at a task (but not process) level.
+Application Skeleton is a simple and powerful tool to generate skeleton applications --- easy-to-program, easy-to-run applications --- that closely mimic a real applications' parallel or distributed performance at a task (but not process) level.
 
-Application classes that can be represented include: bag of tasks, map reduce, multi-stage workflow, and variations of these with a fixed number of iterations.
+It is intended for computer scientists who need to use applications to verify the effectiveness of new system designs, to bypass obstacles that are often encountered when accessing and building real applications. Using the applications generated by Application Skeleton guarantees that the CS systems' effectiveness on synthetic applications will apply to the real applications.
+ 
+Application classes that can be represented include: bag of tasks, map reduce, multi-stage workflow, and variations of these with a fixed number of iterations.  These applications can be generally considered many-task applications.
 
 Applications are described as one or more stages.
 
 Stages are described as one more more tasks.  Stages can also be iterative.
 
 Tasks can be serial or parallel, and have compute and I/O (read and write) elements.
+
+Once created, the applications can be run on single-core, single-node, multi-core, or multi-node (distributed or parallel) computers, depending on what workflow system is used to run them. The generated applications are compatible with workflow system such as Swift and Pegasus, as well as the ubiquitous UNIX shell. The application can also be created as a generic JSON object that can be used by other systems such as the AIMES middleware.
 
 =======
 
