### 2016-05-11 #17 @danielskatz
I've now updated the community guidelines, and we will think about your 'application' comment a bit more, then do something tbd.
