### 2016-05-11 43c4704
update app and user descriptions
