+++ README.md
@@ -4,9 +4,9 @@
 Skeleton
 ========
 
-Application Skeleton is a simple and powerful tool to generate skeleton applications --- easy-to-program, easy-to-run applications --- that closely mimic a real applications' parallel or distributed performance at a task (but not process) level.
+Application Skeleton is a simple and powerful tool to generate skeleton science and engineering applications (for example, modeling and simulation, data analysis) --- easy-to-program, easy-to-run applications --- that closely mimic a real applications' parallel or distributed performance at a task (but not process) level.
 
-It is intended for computer scientists who need to use applications to verify the effectiveness of new system designs, to bypass obstacles that are often encountered when accessing and building real applications. Using the applications generated by Application Skeleton guarantees that the CS systems' effectiveness on synthetic applications will apply to the real applications.
+It is intended for applied computer scientists who need to use science and engineering applications to verify the effectiveness of new systems designed to efficiently run such applications, so that they can bypass obstacles that they often encounter when accessing and building real science and engineering applications. Using the applications generated by Application Skeleton guarantees that the CS systems' effectiveness on synthetic applications will apply to the real applications.
  
 Application classes that can be represented include: bag of tasks, map reduce, multi-stage workflow, and variations of these with a fixed number of iterations.  These applications can be generally considered many-task applications.
 
