### 2016-05-11 81c66c0
update app and user description
