### 2016-05-11 #17 @danielskatz
@krother : I've updated the description of the applications to address your point, and also made the description of the users/need a little more clear, I hope, in both the README and the paper.

Please let us know if this suffices, or if you think it should be further improved.
