### 2016-05-11 #17 @danielskatz
@arfon As we make changes in paper.md and the bib entries, how does the compiled paper get updated?  Or does it not until the review process ends?  It might be nice to have some way that the original submitter could recompile, or a hook that recompiles on changes to specific files in the repo once a submission to JOSS has been made.  If this makes sense, perhaps it should be in issue in the JOSS repo, rather than a comment here?
