### 2016-05-11 #17 @krother
Great, I think the text is sufficiently clear now. I'm ticking all the boxes, let's see what happens next.

I have no idea about how to update the paper. I think the JOSS repo is the right place for that.
