+++ paper.bib
@@ -51,7 +51,7 @@
   publisher = {Springer},
   pages = {131-140},
   volume = {3165},
-  year = {2004}
+  year = {2004},
   doi={10.1007/978-3-540-28642-4_2},
 }
 
