### 2016-05-11 #17 @danielskatz
Please update the paper again, as there have been a couple of typos fixes.  Otherwise, I think this is done. - thanks @krother
