### 2016-05-12 #17 @arfon
This is now accepted and live on JOSS!

- http://joss.theoj.org/papers/10.21105/joss.00017
- http://dx.doi.org/10.21105/joss.00017
