### 2016-05-12 #17 @hausen
Some chars (the curly quotes and the dashes) are garbled in the reference section.
