### 2016-05-12 #17 @arfon
> Some chars (the curly quotes and the dashes) are garbled in the reference section (web version only, the PDF is ok)

Yeah, not sure what's going on there. Will investigate.
