### 2016-05-12 #17 @arfon
> Some chars (the curly quotes and the dashes) are garbled in the reference section (web version only, the PDF is ok)

@hausen this should be fixed now. Can you confirm things are looking good for you now?
