### 2016-05-12 #17 @krother
Hi Arfon,

this is looking good. Great job starting JOSS. I am very happy about the process!

Best regards,
Kristian

----

Dr. Kristian Rother
(+49) 0176 3052 4691
www.academis.eu

Create Better Science!
