### 2016-05-12 #17 @Spencerx
Some clarity through guidelines should avoid doubt when another review has
to be done sooner or later to avoid any doubt or confusion.
